Drobne zmiany i dzialaja od teraz pozycje na doublach

+++ app/src/main/java/com/example/astro/data/Coord.java
@@ -6,23 +6,23 @@ import java.util.Map;
 
 public class Coord {
 
-    private Integer lon;
-    private Integer lat;
+    private Double lon;
+    private Double lat;
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
-    public Integer getLon() {
+    public Double getLon() {
         return lon;
     }
 
-    public void setLon(Integer lon) {
+    public void setLon(Double lon) {
         this.lon = lon;
     }
 
-    public Integer getLat() {
+    public Double getLat() {
         return lat;
     }
 
-    public void setLat(Integer lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
